Produce material based footsteps on civilian movement
Added multiple physical materials for all verieties of sounds including updates
to relavent textures. Created an an animation notification to trigger footsteps
based on the current material under the civilian. This should be re-usable for
the player later.

The bridge currently does not produce sounds, need to take a deeper look
at this. Also added a stub for swimming in the footsteps flow for later.

Made some minor visual and sound mixing adjustments to improve park
ambience.