Editing scene based on performance analysis
After analysing the performance I figured out that most of the frame is
taken up by gpu rendering. This was exhasperated by the excessive
foliage. I added culling and removed about half the foliage improving
the performance. Also added some graphics and vsync settings to allow
the users to adjust these on the fly if they cant support the default
settings.